game backend PoC
