+++ backend/src/game/game.resolver.ts
@@ -1,0 +1,7 @@
+import { Resolver } from '@nestjs/graphql';
+import { GameService } from './game.service';
+
+@Resolver()
+export class GameResolver {
+  constructor(private readonly gameService: GameService) {}
+}
