+++ backend/src/tools/ExceptionFilter.ts
@@ -1,12 +1,16 @@
 import {
   ArgumentsHost,
+  BadRequestException,
   Catch,
   ExceptionFilter,
   HttpException,
 } from '@nestjs/common';
 import { GqlArgumentsHost, GqlExceptionFilter } from '@nestjs/graphql';
+import { BaseWsExceptionFilter, WsException } from '@nestjs/websockets';
 import { ApolloError } from 'apollo-server-express';
-import { TypeORMError } from 'typeorm';
+import { EntityNotFoundError, TypeORMError } from 'typeorm';
+import { TokenExpiredError } from 'jsonwebtoken';
+import { Socket } from 'socket.io';
 
 @Catch(HttpException)
 export class HttpExceptionFilter implements GqlExceptionFilter {
@@ -31,5 +35,22 @@ export class AllExceptionFilter implements ExceptionFilter {
   catch(exception: unknown): any {
     if (exception instanceof ApolloError) throw exception;
     return exception;
+  }
+}
+
+@Catch(WsException, EntityNotFoundError, TokenExpiredError, BadRequestException)
+export class CustomPrcExceptionFilter extends BaseWsExceptionFilter {
+  catch(
+    exception:
+      | WsException
+      | EntityNotFoundError
+      | TokenExpiredError
+      | BadRequestException,
+    host: ArgumentsHost,
+  ) {
+    // For some reason this throws an 'Error' again?
+    //super.catch(exception, host);
+    const client = host.switchToWs().getClient<Socket>();
+    client.emit('exception', { status: 'error', message: exception.message });
   }
 }
