+++ backend/src/tools/UserFromWs.ts
@@ -1,0 +1,16 @@
+import { createParamDecorator, ExecutionContext } from '@nestjs/common';
+import { Socket } from 'socket.io';
+
+// jwt payload soon™
+export const CurrentUserFromWs = createParamDecorator(
+  (data: string, ctx: ExecutionContext) => {
+    const client = ctx.switchToWs().getClient<Socket>();
+    const user = client.data.user;
+
+    if (!user) {
+      return null;
+    }
+
+    return data ? user[data] : user; // extract a specific property only if specified or get a user object
+  },
+);
